+++ .markdownlint-cli2.mjs
@@ -9,10 +9,10 @@ export default {
     },
     MD013: false,
     MD024: {
-      allow_different_nesting: true,
+      siblings_only: true,
     },
     MD033: {
-      allowed_elements: ["style", "script", "AppearanceSwitch", "Badge"],
+      allowed_elements: ["style", "script", "ColorModeSwitch", "Badge"],
     },
     MD035: {
       style: "---",
@@ -24,7 +24,7 @@ export default {
   ignores: [
     "**/node_modules/**",
     "**/__tests__/**",
-    "CHANGELOG.md",
+    "**/CHANGELOG.md",
     "LICENSE",
     "**/*.snippet.md",
   ],