+++ packages/katex-slim/src/plugin.ts
@@ -5,7 +5,11 @@ import { tex } from "@mdit/plugin-tex";
 import type { KatexOptions as OriginalKatexOptions } from "katex";
 import type MarkdownIt from "markdown-it";
 
-import type { KatexToken, MarkdownItKatexOptions } from "./options.js";
+import type {
+  KatexToken,
+  MarkdownItKatexOptions,
+  TeXTransformer,
+} from "./options.js";
 import { escapeHtml } from "./utils.js";
 
 const require = createRequire(import.meta.url);
@@ -23,50 +27,47 @@ try {
 const katexInline = (
   tex: string,
   options: OriginalKatexOptions,
-  vPre: boolean,
+  transformer?: TeXTransformer,
 ): string => {
+  let result: string;
+
   try {
-    const result = katexLib.renderToString(tex, {
+    result = katexLib.renderToString(tex, {
       ...options,
       displayMode: false,
     });
-
-    return vPre
-      ? result.replace(/ class="katex"/g, ' v-pre class="katex"')
-      : result;
   } catch (error) {
     if (options.throwOnError) console.warn(error);
 
-    return `<span ${
-      vPre ? "v-pre " : ""
-    }class='katex-error' title='${escapeHtml(
+    result = `<span class='katex-error' title='${escapeHtml(
       (error as Error).toString(),
     )}'>${escapeHtml(tex)}</span>`;
   }
+
+  return transformer?.(result, false) ?? result;
 };
 
 const katexBlock = (
   tex: string,
   options: OriginalKatexOptions,
-  vPre: boolean,
+  transformer?: TeXTransformer,
 ): string => {
+  let result: string;
+
   try {
-    return `<p ${vPre ? "v-pre " : ""}class='katex-block'>${katexLib.renderToString(
-      tex,
-      {
-        ...options,
-        displayMode: true,
-      },
-    )}</p>\n`;
+    result = `<p class='katex-block'>${katexLib.renderToString(tex, {
+      ...options,
+      displayMode: true,
+    })}</p>\n`;
   } catch (error) {
     if (options.throwOnError) console.warn(error);
 
-    return `<p ${
-      vPre ? "v-pre " : ""
-    }class='katex-block katex-error' title='${escapeHtml(
+    result = `<p class='katex-block katex-error' title='${escapeHtml(
       (error as Error).toString(),
     )}'>${escapeHtml(tex)}</p>\n`;
   }
+
+  return transformer?.(result, true) ?? result;
 };
 
 export const katex = <MarkdownItEnv = unknown>(
@@ -85,7 +86,7 @@ export const katex = <MarkdownItEnv = unknown>(
     mhchem = false,
     logger = (errorCode: string): string =>
       errorCode === "newLineInDisplayMode" ? "ignore" : "warn",
-    vPre = false,
+    transformer,
     ...userOptions
   } = options;
 
@@ -112,8 +113,8 @@ export const katex = <MarkdownItEnv = unknown>(
       };
 
       return displayMode
-        ? katexBlock(content, katexOptions, vPre)
-        : katexInline(content, katexOptions, vPre);
+        ? katexBlock(content, katexOptions, transformer)
+        : katexInline(content, katexOptions, transformer);
     },
   });
 };